requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi
cryptography
Django==4.0.3
djangorestframework==3.13.1
pycparser
PyMySQL
pytz==2022.1
sentry-sdk==1.5.8
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
wincertstore==0.2
